Rename a file ('save_rps_as_csv' -> 'receipt_products_2_csv')
